Fix #38 - Replace Task.number type from byte to short

+++ core/plugins/org.activitymgr.core.dao/src/org/activitymgr/core/dao/TaskDAOCache.java
@@ -37,7 +37,7 @@ public class TaskDAOCache {
 		if (!taskByPathCache.containsKey(fullpath)) {
 			int pathLength = fullpath.length() - Task.PATH_SGM_SIZE;
 			String path = fullpath.substring(0, pathLength);
-			byte number = StringHelper.toByte(fullpath
+			short number = StringHelper.toShort(fullpath
 					.substring(pathLength));
 
 			Task task = selectTaskBySegment(path, "number", number);
